+++ servant-util-beam-pg/CHANGES.md
@@ -2,3 +2,4 @@ Unreleased
 =====
 
 * Add implementation of sorting.
+* Add implementation of pagination.